+++ packages/frontend/src/components/DataViz/store/cartesianChartBaseSlice.ts
@@ -194,14 +194,12 @@ export const cartesianChartConfigSlice = createSlice({
             action: PayloadAction<{
                 reference: string;
                 label: string;
-                index: number;
             }>,
         ) => {
             if (!display) display = {};
             display.series = display.series || {};
             display.series[action.payload.reference] = {
                 ...display.series[action.payload.reference],
-                yAxisIndex: action.payload.index,
                 label: action.payload.label,
             };
         },
@@ -256,23 +254,6 @@ export const cartesianChartConfigSlice = createSlice({
             if (state.display) {
                 if (state.display.yAxis) {
                     state.display.yAxis.splice(index, 1);
-                }
-
-                if (state.display.series) {
-                    Object.entries(state.display.series).forEach(
-                        ([key, series]) => {
-                            if (series.yAxisIndex === index) {
-                                // NOTE: Delete series from display object when it's removed from yAxis
-                                delete state.display?.series?.[key];
-                            } else if (
-                                // NOTE: Decrease yAxisIndex for series that are after the removed yAxis
-                                series.yAxisIndex &&
-                                series.yAxisIndex > index
-                            ) {
-                                series.yAxisIndex--;
-                            }
-                        },
-                    );
                 }
             }
         },
@@ -356,7 +337,6 @@ export const cartesianChartConfigSlice = createSlice({
             display.series[reference] = {
                 ...display.series[reference],
                 format: validFormat,
-                yAxisIndex: index,
             };
 
             if (index === 0) {
@@ -370,7 +350,6 @@ export const cartesianChartConfigSlice = createSlice({
         setSeriesChartType: (
             { display },
             action: PayloadAction<{
-                index: number;
                 type: NonNullable<
                     CartesianChartDisplay['series']
                 >[number]['type'];
@@ -381,17 +360,15 @@ export const cartesianChartConfigSlice = createSlice({
             display = display || {};
             display.series = display.series || {};
 
-            const { index, type, reference } = action.payload;
+            const { type, reference } = action.payload;
             display.series[reference] = {
                 ...display.series[reference],
-                yAxisIndex: index,
                 type,
             };
         },
         setSeriesYAxis: (
             { display },
             action: PayloadAction<{
-                index: number;
                 whichYAxis: NonNullable<
                     CartesianChartDisplay['series']
                 >[number]['whichYAxis'];
@@ -402,10 +379,9 @@ export const cartesianChartConfigSlice = createSlice({
             display = display || {};
             display.series = display.series || {};
 
-            const { index, whichYAxis, reference } = action.payload;
+            const { whichYAxis, reference } = action.payload;
             display.series[reference] = {
                 ...display.series[reference],
-                yAxisIndex: index,
                 whichYAxis,
             };
         },
@@ -414,7 +390,6 @@ export const cartesianChartConfigSlice = createSlice({
             action: PayloadAction<{
                 reference: string;
                 color: string;
-                index?: number;
             }>,
         ) => {
             if (!fieldConfig) return;
@@ -422,13 +397,9 @@ export const cartesianChartConfigSlice = createSlice({
             display.yAxis = display.yAxis || [];
             display.series = display.series || {};
 
-            if (
-                action.payload.index !== undefined &&
-                action.payload.reference !== undefined
-            ) {
+            if (action.payload.reference !== undefined) {
                 display.series[action.payload.reference] = {
                     ...display.series[action.payload.reference],
-                    yAxisIndex: action.payload.index, // TODO: Why set the index here?
                     color: action.payload.color,
                 };
             }
@@ -438,7 +409,6 @@ export const cartesianChartConfigSlice = createSlice({
             action: PayloadAction<{
                 reference: string;
                 valueLabelPosition: ValueLabelPositionOptions;
-                index?: number;
             }>,
         ) => {
             if (!fieldConfig) return;
@@ -446,13 +416,9 @@ export const cartesianChartConfigSlice = createSlice({
             display.yAxis = display.yAxis || [];
             display.series = display.series || {};
 
-            if (
-                action.payload.index !== undefined &&
-                action.payload.reference !== undefined
-            ) {
+            if (action.payload.reference !== undefined) {
                 display.series[action.payload.reference] = {
                     ...display.series[action.payload.reference],
-                    yAxisIndex: action.payload.index, // TODO: Why set the index here?
                     valueLabelPosition: action.payload.valueLabelPosition,
                 };
             }