+++ packages/frontend/src/features/metricsCatalog/components/ExploreMetricButton.tsx
@@ -72,18 +72,13 @@ export const ExploreMetricButton = ({ row }: Props) => {
         >
             <Button
                 compact
-                bg="linear-gradient(180deg, #202B37 0%, #151C24 100%)"
-                radius="md"
+                variant="darkPrimary"
                 onClick={handleExploreClick}
                 py="xxs"
                 px={10}
                 h={32}
                 fz="sm"
                 fw={500}
-                sx={{
-                    border: `1px solid #414E62`,
-                    boxShadow: '0px 0px 0px 1px #151C24',
-                }}
             >
                 Explore
             </Button>